Added rr-delete where U can delete reaction roles made

+++ commands/admin/reset-prefix.js
@@ -4,7 +4,7 @@ const prefix = require('../../config.json').prefix
 module.exports = {
     name : 'reset-prefix',
     run : async(client, message) => {
-       
+        if (!message.member.permissions.has('ADMINISTRATOR')) return;
           
                 await prefixSchema.findOneAndDelete({ Guild : message.guild.id })
                 message.channel.send(`The prefix has been reset to ${prefix}`)
